Added react-router and added routing between 2 pages
