+++ README.md
@@ -12,4 +12,4 @@ Prerequisites for client-side:
 
 After cloning the repository, run `npm install` from the client folder to download the dependancies.
 
-Run `npm install -g webpack` to install webpack globally - you can then build the client project using `webpack` or `npm build`.
+Run `npm install -g webpack@2.1.0-beta.22 --save-dev` to install webpack globally - you can then build the client project using `webpack` or `npm build`.
