+++ client/src/models/LostItem.ts
@@ -1,0 +1,5 @@
+
+ class Item {
+	 Name:string;
+	 Location:string;
+ }
